Fix inline CLI option values
