+++ .changeset/calm-tools-remember.md
@@ -1,0 +1,5 @@
+---
+"effect": patch
+---
+
+Preserve equals signs in inline CLI option values after the first separator.
