--- packages/effect/test/unstable/cli/LexerInlineEquals.audit.test.ts
@@ -1,13 +1,0 @@
-import { assert, it } from "@effect/vitest"
-import * as Lexer from "effect/unstable/cli/internal/lexer"
-
-it("preserves every equals sign in a long option's inline value", () => {
-  const result = Lexer.lex(["--query=left=right"])
-
-  assert.deepStrictEqual(result.tokens, [{
-    _tag: "LongOption",
-    name: "query",
-    raw: "--query=left=right",
-    value: "left=right"
-  }])
-})
